Added success and error messagess to member controller actions

+++ Trident/Controllers/MemberController.cs
@@ -40,22 +40,32 @@ namespace Trident.Controllers
         [HttpPost]
         public ActionResult Create(string MemberName_New, int MemberLevel_New, string MemberSpecialty_New, int MemberStrikes_New, int? MemberTeam_New)
         {
-            //Query string
-            string query = "insert into members (MemberName, MemberLevel, MemberSpecialty, MemberStrikes, team_TeamID) values (@name, @level, @specialty, @strikes, @tid)";
+            if (ModelState.IsValid)
+            {
+                //Query string
+                string query = "insert into members (MemberName, MemberLevel, MemberSpecialty, MemberStrikes, team_TeamID) values (@name, @level, @specialty, @strikes, @tid)";
 
-            //Parameters for the query
-            MySqlParameter[] myParams = new MySqlParameter[5];
-            myParams[0] = new MySqlParameter("@name", MemberName_New);
-            myParams[1] = new MySqlParameter("@level", MemberLevel_New);
-            myParams[2] = new MySqlParameter("@specialty", MemberSpecialty_New);
-            myParams[3] = new MySqlParameter("@strikes", MemberStrikes_New);
-            myParams[4] = new MySqlParameter("@tid", MemberTeam_New);
+                //Parameters for the query
+                MySqlParameter[] myParams = new MySqlParameter[5];
+                myParams[0] = new MySqlParameter("@name", MemberName_New);
+                myParams[1] = new MySqlParameter("@level", MemberLevel_New);
+                myParams[2] = new MySqlParameter("@specialty", MemberSpecialty_New);
+                myParams[3] = new MySqlParameter("@strikes", MemberStrikes_New);
+                myParams[4] = new MySqlParameter("@tid", MemberTeam_New);
 
-            //Execute Query
-            db.Database.ExecuteSqlCommand(query, myParams);
+                //Execute Query
+                db.Database.ExecuteSqlCommand(query, myParams);
 
-            //Re-direct to list of members
-            return RedirectToAction("List");
+                TempData["AddSuccess"] = "Member successfully added";
+                //Re-direct to list of members
+                return RedirectToAction("List");
+            }
+            else
+            {
+                TempData["AddFail"] = "Failed to add member";
+                //Re-direct to list of members
+                return RedirectToAction("List");
+            }
         }
 
         [Authorize(Roles = "Member, Admin")]
@@ -88,43 +98,62 @@ namespace Trident.Controllers
         [HttpPost]
         public ActionResult Edit(int id, string MemberName, int MemberLevel, string MemberSpecialty, int MemberStrikes, int? MemberTeam)
         {
-            if((id == null) || (db.Members.Find(id) == null))
+            if (ModelState.IsValid)
             {
-                return HttpNotFound();
-            }
-            string query = "update members set MemberName=@name, MemberLevel=@level, MemberSpecialty=@specialty, MemberStrikes=@strikes, team_TeamID=@tid where MemberID=@id";
-            MySqlParameter[] myParams = new MySqlParameter[6];
-            myParams[0] = new MySqlParameter("@name", MemberName);
-            myParams[1] = new MySqlParameter("@level", MemberLevel);
-            myParams[2] = new MySqlParameter("@specialty", MemberSpecialty);
-            myParams[3] = new MySqlParameter("@id", id);
-            myParams[4] = new MySqlParameter("@strikes", MemberStrikes);
-            myParams[5] = new MySqlParameter("@tid", MemberTeam);
+                if ((id == null) || (db.Members.Find(id) == null))
+                {
+                    return HttpNotFound();
+                }
+                string query = "update members set MemberName=@name, MemberLevel=@level, MemberSpecialty=@specialty, MemberStrikes=@strikes, team_TeamID=@tid where MemberID=@id";
+                MySqlParameter[] myParams = new MySqlParameter[6];
+                myParams[0] = new MySqlParameter("@name", MemberName);
+                myParams[1] = new MySqlParameter("@level", MemberLevel);
+                myParams[2] = new MySqlParameter("@specialty", MemberSpecialty);
+                myParams[3] = new MySqlParameter("@id", id);
+                myParams[4] = new MySqlParameter("@strikes", MemberStrikes);
+                myParams[5] = new MySqlParameter("@tid", MemberTeam);
 
-            db.Database.ExecuteSqlCommand(query, myParams);
-            return RedirectToAction("Show/" + id);
+                db.Database.ExecuteSqlCommand(query, myParams);
+                TempData["EditSuccess"] = "Member successfully edited";
+                return RedirectToAction("Show/" + id);
+            }
+            else
+            {
+                TempData["EditFail"] = "Failed to edit member";
+                return RedirectToAction("Show/" + id);
+            }
         }
 
         [Authorize(Roles = "Admin")]
         public ActionResult Delete(int? id)
         {
-            if ((id == null) || (db.Members.Find(id) == null))
+            if(ModelState.IsValid)
             {
-                return HttpNotFound();
-            }
-            string query;
-            MySqlParameter param = new MySqlParameter("@id", id);
+                if ((id == null) || (db.Members.Find(id) == null))
+                {
+                    return HttpNotFound();
+                }
+                string query;
+                MySqlParameter param = new MySqlParameter("@id", id);
 
-            //Delete associated characters
-            query = "delete from characters where member_MemberID=@id";
-            param = new MySqlParameter("@id", id);
-            db.Database.ExecuteSqlCommand(query, param);
+                //Delete associated characters
+                query = "delete from characters where member_MemberID=@id";
+                param = new MySqlParameter("@id", id);
+                db.Database.ExecuteSqlCommand(query, param);
             
-            //Delete member
-            query = "delete from members where MemberID=@id";
-            param = new MySqlParameter("@id", id);
-            db.Database.ExecuteSqlCommand(query, param);
-            return RedirectToAction("List");
+                //Delete member
+                query = "delete from members where MemberID=@id";
+                param = new MySqlParameter("@id", id);
+                db.Database.ExecuteSqlCommand(query, param);
+
+                TempData["DeleteSuccess"] = "Member successfully deleted";
+                return RedirectToAction("List");
+            }
+            else
+            {
+                TempData["DeleteFail"] = "Failed to delete member";
+                return RedirectToAction("List");
+            }
 
         }
